Fixed some file size issues.
git-svn-id: https://sirius.chem.vt.edu/svn/psi4/trunk@1343 6a214465-e6f0-41bc-91e6-de896bb96b5e

Former-commit-id: 5e1657a5c479683710ac50b8cd16221fcbaa44a4

+++ src/lib/libsapt_solver/disp21.cc
@@ -63,8 +63,8 @@ double SAPT2p::disp210()
   double **tARBS = read_IJKL(PSIF_SAPT_AMPS,"T ARBS Amplitudes",
     calc_info_.noccA*calc_info_.nvirA,calc_info_.noccB*calc_info_.nvirB);
 
-  energy += 4.0*C_DDOT(calc_info_.noccA*calc_info_.nvirA*calc_info_.noccB*
-    calc_info_.nvirB,tARBS[0],1,gARBS[0],1);
+  energy += 4.0*C_DDOT((long int) calc_info_.noccA*calc_info_.nvirA*
+    calc_info_.noccB*calc_info_.nvirB,tARBS[0],1,gARBS[0],1);
 
   free_block(gARBS);
   free_block(tARBS);
@@ -93,8 +93,8 @@ double SAPT2p::disp201()
   double **tARBS = read_IJKL(PSIF_SAPT_AMPS,"T ARBS Amplitudes",
     calc_info_.noccA*calc_info_.nvirA,calc_info_.noccB*calc_info_.nvirB);
 
-  energy += 4.0*C_DDOT(calc_info_.noccA*calc_info_.nvirA*calc_info_.noccB*
-    calc_info_.nvirB,tARBS[0],1,gARBS[0],1);
+  energy += 4.0*C_DDOT((long int) calc_info_.noccA*calc_info_.nvirA*
+    calc_info_.noccB*calc_info_.nvirB,tARBS[0],1,gARBS[0],1);
   
   free_block(gARBS);
   free_block(tARBS);
